refactor file structure to be more consistent
